requirements.txt
typing_extensions
PyGLM
PyOpenGL
PyOpenGL_accelerate
numpy==1.26.4